2025-03-01 Roman (stable) - !!!:LibsDB.OscadaLibs:     !!: Appended of the user "RemCntr" with password "RemCntr" and in the group DAQ for use in         the initiative remote connections to the OpenSCADA server.     !!: Appended of the disabled transports "RemCntr" to Transport.Sockets and Transport.SSL,         that is when you want to provide data of your station, you just enable one of them.     !!!:tmplib_base.initRemCntr: Added of access control on the remote connection and blocking             the not permitted host during configurable time.         !!!: Added of removing dead connections during configurable time.         General improving in saving and processing errors.         Added of limiting size of the remote host ID in 20 symbols and description.     tmplib_DevLib.ergomera125: Added of description.     LowDevLib.{S1BP,1W}: Translated to Ukrainian and added of description. - !:LibsDB.vcaBase: wlb_Main.{grph_panel,doc_panel}: Adding of checking for activity the controlled     trend or document, and closing at its inactivity. - !!!!:SYS,DAQ,UI.VCAEngine: Implementing of disabling translation of string configuration fields,         DAQ attributes, UI attributes at changing in the dynamic execution:     - SYS,DAQ,DB.{FireBird,MySQL,ODBC,PostgreSQL,SQLite}: appending the unified method {TCfg,TVal}::isTransl()         of checking translation the element, and moving on its everywhere;     - TVal: appending the "mNoTransl" flag and functions of its treating:         noTransl(), setNoTransl(), TValue::setNoTransl();     - UI.VCAEngine::Attr: appending the "NoTransl" flag and its treating in isTransl();     - translation is disabled:         - at writing through links of implementation the templates in TPrmTempl::Impl::lnkOutput();         - at the User API call TVal::set();         - UI.VCAEngine: at Widget attribute change in the Session context for widgets in processing             and at writing to DAQ attribute by links.     - the DAQ Parameters added of restoring the translation of dynamically changed configuration fields at load. - !!:FIX:Transports,Protocols: Unification of errors generation in calls messProtIO() and its User API variant             messIO(XMLNodeObj req, string prt) in accordance the documentation,             that is obligatory with the error code:         - adding en error code Tr_ErrUnknown for unknown transport side errors;         - Protocol.{ModBus,HTTP,SelfSystem,UserProtocol}: adapted-updated to the errors generation with the error code. - !!:DAQ.{LogicLev,ModBus,OPC_UA,Siemens},UI.{QTCfg,WebCfgD}: Implementing of ordering value of the updating time         of the configuration interface of the calculated objects, to update after the object processing:     - time of the ordered update is passed in seconds through the "updTm" attribute in result to the "set" commands         of fields;     - implemented for the tab "Attributes (/val)" and "Parameter templates (/tmpl)" of the Logical Based sources         and in the periodic mode. - FIX:TSecurity,UI.VCAEngine: Fixed in loading non initiated (without storage) users and groups at force reloading. - FIX:Archive.FSArch,Transport.{Sockets,SSL}: Fixed in building for Android. - ModelsDB/{AGLKS/oscada_AGLKS.xml,Boiler/oscada_Boiler.xml}: Removing system groups without users,     than allow the groups specify in libraries. - FIX:SYS: Preventing any messages generation during processing events: SIGFPE, SIGABRT, SIGSEGV.     That is only call the crashhandler and exit. - !:FIX:DAQ: The TPrmTempl object is set to stop at changing the procedure text also after loading.     !:FIX: Fixing-finishing the dynamic translation of values requesting in the Control Request         TValue::"get:/serv/attr" for non archiving elements.     TVal: fld() is set to constant.     !:FIX:System: Fixed in the CPU DA about limiting in time of obtaining the CPU load by reading the counters as uint32         and storing previous in int32, that is the counters are both read and stored now in uint64.     !:FIX:DAQGate: Preventing of marking the translation of the string fields, since there is no sense. - !!!:FIX:Transports: Fixing of clearing the connection properties in {TTransportIn,TTransportOut}::clearConPrm().     !!!:FIX:SSL: Fixed in processing responses of the output transports from initiative connections,         especially for no response. The true code got from Sockets with some improving for tries.     FIX:Sockets: Some improved for tries in the code of processing responses of the output transports from initiative         connections, especially for no response. - !!:UI.WebCfgD: Expanding of call pageRefresh() by integrating the lag time.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@3012 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/System/da_cpu.cpp
@@ -1,7 +1,7 @@
 
 //OpenSCADA module DAQ.System file: da_cpu.cpp
 /***************************************************************************
- *   Copyright (C) 2005-2024 by Roman Savochenko, <roman@oscada.org>       *
+ *   Copyright (C) 2005-2025 by Roman Savochenko, <roman@oscada.org>       *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -57,11 +57,11 @@ void CPU::updGen( bool cntr )
 
 	// Updating
 	string resLoad, resSys, resUser, resIdle;
-	long unsigned user, nice, sys, idle, iowait;
+	long long unsigned user, nice, sys, idle, iowait;
 	char buf[256];
 	for(int n = 0, unit = -1; fgets(buf,sizeof(buf),f) != NULL; ) {
-	    if(!isdigit(buf[3]) && (n=sscanf(buf,"cpu %lu %lu %lu %lu %lu\n",&user,&nice,&sys,&idle,&iowait))) unit = -1;
-	    else if((n=sscanf(buf,"cpu%d %lu %lu %lu %lu %lu\n",&unit,&user,&nice,&sys,&idle,&iowait))) ;
+	    if(!isdigit(buf[3]) && (n=sscanf(buf,"cpu %llu %llu %llu %llu %llu\n",&user,&nice,&sys,&idle,&iowait))) unit = -1;
+	    else if((n=sscanf(buf,"cpu%d %llu %llu %llu %llu %llu\n",&unit,&user,&nice,&sys,&idle,&iowait))) ;
 	    else continue;
 
 	    while(load.size() < (unit+2)) load.push_back(tval());
